% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cci_rank.R
\name{add_cci_rank}
\alias{add_cci_rank}
\title{Add CCI rank}
\usage{
add_cci_rank(df, n_perm = 1000)
}
\arguments{
\item{df}{dataframe with at least the columns pval and interaction_score}

\item{n_perm}{no. permutations to define smallest possible pvalue}
}
\value{
df with columns pval_corr, log10_score and rank
}
\description{
Add cci rank based on p-value x interaction score. Needed for CCI methods that do not provide a rank, i.e. CellChat, CellPhoneDB and cell2cell.
}
