% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_in_multi_samples.R
\name{add_detection_flags}
\alias{add_detection_flags}
\title{Add detection flags}
\usage{
add_detection_flags(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df dataframe
}
\description{
adds columns indicating how many patients/samples detected a source-target x interaction combination, within a voting strategy and condition/group. Also adds columns with the exact patients/samples where the combination was found.
}
