% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByDetectionInMultiSamples.R
\name{AddDetectionFlags}
\alias{AddDetectionFlags}
\title{Add detection flags}
\usage{
AddDetectionFlags(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df dataframe
}
\description{
adds columns indicating how many patients/samples detected a source-target x interaction combination, within a voting strategy and condition/group. Also adds columns with the exact patients/samples where the combination was found.
}
