% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetadata.R
\name{AddMetaData}
\alias{AddMetaData}
\title{Combine samples}
\usage{
AddMetaData(
  df,
  meta_df,
  patient_var = "sample_id",
  condition_var = "condition_dummy",
  sample_var = "sample_id"
)
}
\arguments{
\item{df}{dataframe}

\item{patient_var}{patient variable name in metadata}

\item{condition_var}{condition variable name in metadata (default = "Condition_dummy")}

\item{sample_var}{sample variable name in metadata (default = "Sample")}

\item{metadata}{dataframe with metadata}
}
\description{
hard-combine results per sample.
}
