% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{AddProteins}
\alias{AddProteins}
\title{Add proteins}
\usage{
AddProteins(vec, gene_uniprot_table)
}
\arguments{
\item{vec}{vector}

\item{gene_uniprot_table}{dataframe containing at least 'genesymbol' and 'uniprot' columns}
}
\value{
string with all protein complexes that are involved separate by an underscore, if no protein found then returns vector of NAs (same length as input vector)
}
\description{
Add proteins
}
