% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUID.R
\name{AddUniqueIdentifier}
\alias{AddUniqueIdentifier}
\title{Add unique identifier}
\usage{
AddUniqueIdentifier(
  df,
  cols_to_merge = c("sample_id", "source_target", "complex_interaction")
)
}
\arguments{
\item{df}{dataframe}

\item{cols_to_merge}{columns that should be used to create the uid}
}
\value{
df
}
\description{
Add unique identifier
}
