% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\name{AggregateCCIRanks}
\alias{AggregateCCIRanks}
\title{RRA interactions}
\usage{
AggregateCCIRanks(list_of_cci_objects, n_perm = 1000)
}
\arguments{
\item{list_of_cci_objects}{named list of CCI objects, should contain: CellChatv2, LIANA, Cell2Cell, CellPhoneDBv5}

\item{n_perm}{number of permutations to create pval_dummy variable for computing -log10(pval_dummy) * score, only applicable to CellChat, Cell2Cell and CellPhoneDB}
}
\value{
dataframe
}
\description{
Rank interactions based on CellChat, LIANA (use all methods), cell2cell and CellPhoneDB
}
