% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSamples.R
\name{AggregateSamples}
\alias{AggregateSamples}
\title{Aggregate samples w/ ranked interactions}
\usage{
AggregateSamples(df, method = "BH")
}
\arguments{
\item{df}{dataframe with interaction results for all samples (samples_interactions_agg_rank.rds)}

\item{method}{correction method}
}
\value{
dataframe
}
\description{
Combine p-values/scores by condition
}
