% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{CheckLengthGeneSymbol}
\alias{CheckLengthGeneSymbol}
\title{Check genesymbol length}
\usage{
CheckLengthGeneSymbol(vec)
}
\arguments{
\item{vec}{vector with genesymbols}
}
\description{
Check genesymbol length
}
