% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_samples.R
\name{combine_samples}
\alias{combine_samples}
\title{Combine samples}
\usage{
combine_samples(
  df,
  metadata,
  patient_var = "sample_id",
  condition_var = "condition_dummy",
  sample_var = "sample_id"
)
}
\arguments{
\item{metadata}{path to file with corresponding metadata (RDS)}

\item{patient_var}{patient variable name in metadata}

\item{condition_var}{condition variable name in metadata (default = "Condition_dummy")}

\item{sample_var}{sample variable name in metadata (default = "Sample")}

\item{input_dir}{df}
}
\description{
hard-combine results per sample.
}
