% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{CreateDB}
\alias{CreateDB}
\title{Create interactions databasee}
\usage{
CreateDB(source_cpdb_dir, output_dir)
}
\arguments{
\item{source_cpdb_dir}{path to 'interaction_input.csv' from CellPhoneDB database}

\item{output_dir}{path to directory for saving the generated CellPhoneDB files}
}
\description{
Generates database for LIANA, CellChat, Cell2Cell and CellPhoneDB
}
