% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\name{extract_df_with_ranks}
\alias{extract_df_with_ranks}
\title{Extract dataframe with ranks}
\usage{
extract_df_with_ranks(df, n_perm = 1000)
}
\arguments{
\item{df}{dataframe with interactions}
}
\value{
dataframe
}
\description{
adds uid + ranks interactions and returns a dataframe with the required columns for downstream analyses
}
