% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_agg_res.R
\name{filter_agg_res}
\alias{filter_agg_res}
\title{Filter agg res by user constraints}
\usage{
filter_agg_res(
  condition_var = "Condition_dummy",
  input_dir = "",
  interactions_agg_binarized = "",
  interactions_agg_continuous = "",
  output_dir = "."
)
}
\arguments{
\item{condition_var}{Variabile containing the 'condition' (group/category) of the samples (default = "Condition_dummy")}

\item{input_dir}{directory with '402a_filtering_detect_in_multi_samples.rds' and '402b_aggregation_samples.rds', alternatively specify path separately ('interactions_agg_binarized' & 'interactions_agg_continuous') (default = "")}

\item{interactions_agg_binarized}{path to '402a_filtering_detect_in_multi_samples.rds' (default = "")}

\item{interactions_agg_continuous}{path to '402b_aggregation_samples.rds' (default = "")}

\item{output_dir}{output directory for saving output (default = '.')}
}
\description{
Filter agg res by user constraints
}
