% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_in_multi_samples.R
\name{filter_by_detection_in_multi_samples}
\alias{filter_by_detection_in_multi_samples}
\title{Filter by detection in multiple samples}
\usage{
filter_by_detection_in_multi_samples(df, min_patients = 2)
}
\arguments{
\item{df}{dataframe from 'samples_interactions_mvoted.rds'}

\item{min_patients}{Minimum number of patients for an interaction to be kept (default = 2)}
}
\value{
dataframe
}
\description{
Filtering detected interactions by looking at recurrence in multiple samples/patients. Keeping source-target x interaction pairs that were found in multiple patients (default >2 ) within a condition/group. Performing filtering for the two voting strategies 'lenient' and 'stringent' (based on detection in multiple tools)
}
