% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDB.R
\name{FilterDB}
\alias{FilterDB}
\title{Filter and clean-up database}
\usage{
FilterDB(db)
}
\arguments{
\item{db}{dataframe with database of interactions, the following columns should at least be present 'source_genesymbol' and 'target_genesymbol'}
}
\description{
Filter and clean-up database
}
