% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cell2cell.R
\name{format_cell2cell}
\alias{format_cell2cell}
\title{Format cell2cell CCI results}
\usage{
format_cell2cell(c2c_df, mode = c("pval", "interaction_score")[1])
}
\arguments{
\item{c2c_df}{dataframe with the inferred interactions from cell2cell (output file generated by Python/202_cci_cell2cell.py, a csv file)}

\item{mode}{'pval' or 'interaction_score' dependent on input}
}
\value{
dataframe with columns: interaction, source_target, pval
}
\description{
Format cell2cell CCI results
}
