% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cellchat.R
\name{format_cellchat}
\alias{format_cellchat}
\title{Format CellChat CCI results}
\usage{
format_cellchat(cc_obj, ref_db, sample_id = NA, n_cores = 1)
}
\arguments{
\item{cc_obj}{CellChat outputs}

\item{ref_db}{dataframe with reference database of interactions}

\item{sample_id}{character string for sample_id to add to dataframe (if non given set to NA)}

\item{n_cores}{integer indicating the number of cores to use}
}
\value{
dataframe with columns: source_target, interaction_score, pval, complex_interaction, method, sample_id
}
\description{
Takes that 'raw' output from CellChat reshapes and reformats that data for downstream aggregation of the CCI results of the different tools.
}
