% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cpdb.R
\name{format_cpdb_wrapper}
\alias{format_cpdb_wrapper}
\title{Format CellPhoneDB v5 results}
\usage{
format_cpdb_wrapper(cpdb_dfs, ref_db, sample_id = NA)
}
\arguments{
\item{cpdb_dfs}{named list of dfs (interaction_scores = 'statistical_analysis_interaction_scores', pval = 'statistical_analysis_pvalues', means = 'statistical_analysis_means, sign_mean = 'statistical_analysis_significant_means')}

\item{ref_db}{dataframe with reference database of interactions}

\item{sample_id}{character string for sample_id to add to dataframe (if non given set to NA)}
}
\value{
dataframe with columns: source_target, interaction_score, pval, rank, sign_mean, mean, method, sample_id, complex_interaction
}
\description{
Format CellPhoneDB v5 results
}
