% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\name{format_internal_methods_liana}
\alias{format_internal_methods_liana}
\title{Format rankings of LIANA's internal methods}
\usage{
format_internal_methods_liana(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with columns: uid, method and rank (larger value = 'lower importance/priority')
}
\description{
Format rankings of LIANA's internal methods
}
