% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_liana.R
\name{format_liana}
\alias{format_liana}
\title{Format LIANA CCI results}
\usage{
format_liana(cci_obj, ref_db, sample_id = NA)
}
\arguments{
\item{cci_obj}{LIANA object with the inferred interactions, generated by run_LIANA()}

\item{ref_db}{dataframe with reference database of interactions}

\item{sample_id}{character string for sample_id to add to dataframe (if non given set to NA)}
}
\value{
dataframe
}
\description{
Format LIANA CCI results
}
