% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cell2cell.R
\name{FormatCell2CellWrapper}
\alias{FormatCell2CellWrapper}
\title{Format cell2cell CCI results}
\usage{
FormatCell2CellWrapper(
  input_interactions_pval,
  input_interactions_scores,
  ref_db,
  sample_id = NA
)
}
\arguments{
\item{input_interactions_pval}{dataframe with the inferred interactions pvalues from cell2cell (output file generated by Python/202_cci_cell2cell.py, a csv file)}

\item{input_interactions_scores}{dataframe with the inferred interactions pvalues from cell2cell (output file generated by Python/202_cci_cell2cell.py, a csv file)}

\item{ref_db}{dataframe with reference database of interactions}

\item{sample_id}{character string for sample_id to add to dataframe (if non given set to NA)}
}
\value{
dataframe with columns: source_target, pval, interaction_score, complex_interaction, method, sample_id
}
\description{
Format cell2cell CCI results
}
