% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cpdb.R
\name{FormatCPDB}
\alias{FormatCPDB}
\title{Format individual output from CellPhoneDB v5}
\usage{
FormatCPDB(df, mode = c("pval", "sign_mean", "mean", "interaction_score")[1])
}
\arguments{
\item{df}{dataframe}

\item{mode}{character string indicating what information the dataframe contains.}
}
\value{
dataframe with columns: interacting_pair, source_target and \if{html}{\out{<mode>}}
}
\description{
Format individual output from CellPhoneDB v5
}
