% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetadata.R
\name{FormatMetadata}
\alias{FormatMetadata}
\title{Format metadata}
\usage{
FormatMetadata(
  meta_df,
  patient_var = "sample_id",
  condition_var = "condition_dummy",
  sample_var = "sample_id"
)
}
\arguments{
\item{meta_df}{dataframe with metadata}

\item{patient_var}{patient variable name in metadata}

\item{condition_var}{condition variable name in metadata (default = "Condition_dummy")}

\item{sample_var}{sample variable name in metadata (default = "Sample")}
}
\value{
dataframe with sample_id, condition, patient_id
}
\description{
format metadata.
}
