% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(input_file, output_dir = ".")
}
\arguments{
\item{input_file}{path to seurat object (rds file)}

\item{output_dir}{output directory for saving output (default = '.')}
}
\description{
Extracts the metadata and saves it as and '.rds' and '.csv' file in user-specified output directory.
}
\examples{
\dontrun{
filepath <- "/johndoe/data/seurat_obj.rds"
output_dir <- "/johndoe/output"
get_metadata(filepath, output_dir)
}
}
