% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_proteins_wrapper}
\alias{get_proteins_wrapper}
\title{get proteins from  OmniPathR based on list of genes}
\usage{
get_proteins_wrapper(genes_df)
}
\arguments{
\item{genes_df}{dataframe with a single column called 'genesymbol'}
}
\description{
get proteins from  OmniPathR based on list of genes
}
