% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{GetProteinsWrapper}
\alias{GetProteinsWrapper}
\title{get proteins from  OmniPathR based on list of genes}
\usage{
GetProteinsWrapper(genes_df)
}
\arguments{
\item{genes_df}{dataframe with a single column called 'genesymbol'}
}
\description{
get proteins from  OmniPathR based on list of genes
}
