% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardcombineDB.R
\name{HardCombineDBs}
\alias{HardCombineDBs}
\title{Hard combine databases}
\usage{
HardCombineDBs(cellchat_db, cpdb_db, liana_db)
}
\arguments{
\item{cellchat_db}{cellchat database}

\item{cpdb_db}{cpdb_database}

\item{liana_db}{liana database}
}
\description{
Combine databases based on 'source_genesymbol' and 'target_genesymbol'
}
