% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{IsMapped}
\alias{IsMapped}
\title{Check mapping}
\usage{
IsMapped(vec, lookup_table, lookup_var = "genesymbol")
}
\arguments{
\item{vec}{vector}

\item{lookup_table}{dataframe for looking up valuees}

\item{lookup_var}{variable in lookup_table for checking overlap with vec}
}
\value{
returns whether all entries of the vector where found in the lookup_table (boolean)
}
\description{
Check whether all entries in vector are found in a dataframe
}
