% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_liana.R
\name{load_liana_dbs}
\alias{load_liana_dbs}
\title{Load LIANA databases}
\usage{
load_liana_dbs()
}
\value{
dataframe with the two databases
}
\description{
Loads and combines the LIANA Consensus and Ramilowski 2015 database (hard combine, missing columns set to NA)
}
