% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCellChatDB.R
\name{LoadCellChatDB}
\alias{LoadCellChatDB}
\title{Load CellChat DB}
\usage{
LoadCellChatDB()
}
\value{
dataframe with 'source_genesymbol' and 'target_genesymbol'
}
\description{
load cellcat database and format for merging
}
