% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCPDBDB.R
\name{LoadCPDBDB}
\alias{LoadCPDBDB}
\title{Load CellPhoneDB DB}
\usage{
LoadCPDBDB(source_cpdb_dir)
}
\arguments{
\item{source_cpdb_dir}{path to 'interaction_input.csv' from CellPhoneDB database}
}
\value{
dataframe with 'source_genesymbol' and 'target_genesymbol'
}
\description{
load CellPhoneDB database and format for merging
}
