% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLIANADBs.R
\name{LoadLIANADBs}
\alias{LoadLIANADBs}
\title{Load LIANA databases}
\usage{
LoadLIANADBs()
}
\value{
dataframe with the two databases
}
\description{
Loads and combines the LIANA Consensus and Ramilowski 2015 database (hard combine, missing columns set to NA)
}
