% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{LookupUniprot}
\alias{LookupUniprot}
\title{Look up uniprot based on genesymbol}
\usage{
LookupUniprot(gene, gene_uniprot_table)
}
\arguments{
\item{gene}{genesymbol}

\item{gene_uniprot_table}{dataframe containing at least 'genesymbol' and 'uniprot' columns}
}
\description{
Look up uniprot based on genesymbol
}
