% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{make_complex}
\alias{make_complex}
\title{make complex}
\usage{
make_complex(vec, collapse = collapse)
}
\arguments{
\item{vec}{vector}

\item{collapse}{character string to separate items default = ':'}
}
\description{
Merge items in vector by collapse character specified by user, default = ':'
}
