% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{make_complex_wrapper}
\alias{make_complex_wrapper}
\title{make complex wrapper}
\usage{
make_complex_wrapper(x, collapse = ":")
}
\arguments{
\item{collapse}{character string to separate items default = ':'}

\item{vec}{vector}
}
\description{
(1) orders vector (alphabetically, NAs last), (2) combines into single string
}
