% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{MakeComplex}
\alias{MakeComplex}
\title{make complex}
\usage{
MakeComplex(vec, collapse = collapse)
}
\arguments{
\item{vec}{vector}

\item{collapse}{character string to separate items default = ':'}
}
\description{
Merge items in vector by collapse character specified by user, default = ':'
}
