% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDB.R
\name{MakeComplex_wrapper}
\alias{MakeComplex_wrapper}
\title{make complex wrapper}
\usage{
MakeComplex_wrapper(x, collapse = ":")
}
\arguments{
\item{collapse}{character string to separate items default = ':'}

\item{vec}{vector}
}
\description{
(1) orders vector (alphabetically, NAs last), (2) combines into single string
}
