% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data for CCI inference}
\usage{
prepareData(seurat_obj, annot, min_cells = 5)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{annot}{variable in metadata containing the cell annotation}

\item{min_cells}{Minimum number of cells required in each cell group for cell-cell communication (default = 5)}
}
\value{
Seurat object
}
\description{
Checks if there are enough cell types with at least \code{min_cells}, then continues with normalization. Groups (cell types) with insufficient no. cells are removed.
}
