% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_seurat_obj}
\alias{reduce_seurat_obj}
\title{Reduce seurat object size}
\usage{
reduce_seurat_obj(input_file, output_dir = ".")
}
\arguments{
\item{input_file}{path to seurat object (rds file)}

\item{output_dir}{output directory for saving output (default = '.')}
}
\description{
remove redundant layers or assays, i.e. only keeping the 'RNA' assay
}
