% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unused_assays.R
\name{remove_unused_assays}
\alias{remove_unused_assays}
\title{Reduce seurat object size}
\usage{
remove_unused_assays(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\description{
remove redundant layers or assays, i.e. only keeping the 'RNA' assay
}
