% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeGroupsWithLowNCells.R
\name{removeGroupsWithLowNCells}
\alias{removeGroupsWithLowNCells}
\title{Filtering the Seurat object to reduce size of object}
\usage{
removeGroupsWithLowNCells(seurat_obj, annot, min_cells = 5)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{annot}{Annotation to use for filtering}

\item{min_cells}{Minimum number of cells per annotation (default = 5)}
}
\value{
seurat_obj filtered seurat object
}
\description{
Only keep cell types with at least N cells (user-defined)
}
\examples{
dontrun{seurat_obj <- filtering(seurat_obj, "custom_annot", 300, genes_oi)}
}
