% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeUnusedAssays.R
\name{RemoveUnusedAssays}
\alias{RemoveUnusedAssays}
\title{Reduce seurat object size}
\usage{
RemoveUnusedAssays(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\description{
remove redundant layers or assays, i.e. only keeping the 'RNA' assay
}
