% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cellchat.R
\name{ReshapeAndAddInteractionName}
\alias{ReshapeAndAddInteractionName}
\title{Reshape & add interaction name}
\usage{
ReshapeAndAddInteractionName(
  interaction_name,
  cci_ls,
  mode = c("pval", "interaction_score")[1]
)
}
\arguments{
\item{interaction_name}{name of interaction generally in format: \if{html}{\out{<ligand>}}__\if{html}{\out{<receptor>}}}

\item{cci_ls}{list with 2D arrays representing the results for the interactions in the (custom) CellChat database}

\item{mode}{character string which CellChat output slot is formatted, i.e. 'interaction_score' or 'pval'}
}
\value{
dataframe with 4 columns, i.e. source, target, interaction_score, interaction
}
\description{
Takes the CellChat a single 2D array representing the probabilities/pvalues for an interaction in the CCI database from the cellchat object, i.e. \code{cc_obj@net$prob} or \code{cc_obj@net$pval} after running CellChat. The array (rownames=source cell type, columnnames=target cell type) is converted to the long-format.
}
