% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellChat.R
\name{RunCellChat}
\alias{RunCellChat}
\title{Run CellChat}
\usage{
RunCellChat(
  gene_expr_path,
  interactions_db_path,
  annot,
  min_cells = 5,
  n_perm = 1000
)
}
\arguments{
\item{gene_expr_path}{seurat object with gene expression (rds file)}

\item{interactions_db_path}{path to (custom) CellChat database which should be an '.rds' file.}

\item{annot}{Column in metadata containing the cell type labels}

\item{min_cells}{Minimum number of cells required in each cell group for cell-cell communication (default = 5)}

\item{n_perm}{Number of permutations for permutation testing (default = 1000)}
}
\value{
cellchat object
}
\description{
Uses a Seurat object to run CellChat, follows the standard CellChat workflow.
}
