% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xlsx.R
\name{save_as_xlsx}
\alias{save_as_xlsx}
\title{Save interactions as xlsx}
\usage{
save_as_xlsx(
  interactions_agg_integration,
  alpha = 0.05,
  output_dir = ".",
  output_name = "interactions_summary",
  is_stringent = FALSE
)
}
\arguments{
\item{interactions_agg_integration}{path to '402c_filtering_aggregated_res.rds'}

\item{alpha}{significance threshold for filtering p-values (default = 0.05)}

\item{output_dir}{output directory for saving output (default = '.')}

\item{output_name}{name for output file (default = "interactions_summary")}

\item{is_stringent}{only keep stringent interactions (default = FALSE)}

\item{condition_var}{Variabile containing the 'condition' (group/category) of the samples (default = "Condition_dummy")}
}
\description{
Save interactions as xlsx
}
