% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_seurat_into_samples}
\alias{split_seurat_into_samples}
\title{Split seurat object by variable}
\usage{
split_seurat_into_samples(input_file, output_dir = ".", sample_var = "Sample")
}
\arguments{
\item{input_file}{path to seurat object (rds file)}

\item{output_dir}{output directory for saving output (default = '.')}

\item{sample_var}{variable to split the object by, must be part of metadata (default = "Sample")}
}
\description{
Splits seurat object by a user-specified variable, e.g. Sample
}
