% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TakeConsensus.R
\name{TakeConsensus}
\alias{TakeConsensus}
\title{Take consensus}
\usage{
TakeConsensus(interactions_df, alpha = 0.05)
}
\arguments{
\item{alpha}{significance threshold (default = 0.05)}

\item{list_of_obj}{list with the dataframes for CellChat, LIANA, cell2cell and cellphonedb}
}
\description{
Assess number of methods that detect an interaction in a source-target
}
