% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wide_cci_df.R
\name{to_wide_cci_df}
\alias{to_wide_cci_df}
\title{Convert to side-by-side comparison format}
\usage{
to_wide_cci_df(df)
}
\arguments{
\item{df}{should contain the columns: method, interaction_score and pval}
}
\value{
dataframe in wide-format interaction_score_\if{html}{\out{<method>}} and pval_\if{html}{\out{<method>}}
}
\description{
Convert to side-by-side comparison format
}
