% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToWideCCIDF.R, R/ToWideCCIDf.R
\name{ToWideCCIDF}
\alias{ToWideCCIDF}
\title{Convert to side-by-side comparison format}
\usage{
ToWideCCIDF(df)

ToWideCCIDF(df)
}
\arguments{
\item{df}{should contain the columns: method, interaction_score and pval}
}
\value{
dataframe in wide-format interaction_score_\if{html}{\out{<method>}} and pval_\if{html}{\out{<method>}}

dataframe in wide-format interaction_score_\if{html}{\out{<method>}} and pval_\if{html}{\out{<method>}}
}
\description{
Convert to side-by-side comparison format

Convert to side-by-side comparison format
}
