% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_cellchat.R
\name{update_cellchat_db}
\alias{update_cellchat_db}
\title{Update CellChat DB using (updated) CellPhoneDB}
\usage{
update_cellchat_db(geneInfo, db, output_dir)
}
\arguments{
\item{geneInfo}{dataframe object containing contents from gene_input.csv from CellPhoneDB}

\item{db}{database}

\item{output_dir}{path to directory for saving the generated CellPhoneDB database (.rds)}
}
\description{
Update CellChat DB using (updated) CellPhoneDB
}
