% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCellChatDB.R
\name{UpdateCellChatDB}
\alias{UpdateCellChatDB}
\title{Update CellChat DB using (updated) CellPhoneDB}
\usage{
UpdateCellChatDB(geneInfo, db, output_dir)
}
\arguments{
\item{geneInfo}{dataframe object containing contents from gene_input.csv from CellPhoneDB}

\item{db}{database}

\item{output_dir}{path to directory for saving the generated CellPhoneDB database (.rds)}
}
\description{
Update CellChat DB using (updated) CellPhoneDB
}
