% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_samples.R
\name{aggregate_samples}
\alias{aggregate_samples}
\title{Aggregate samples w/ ranked interactions}
\usage{
aggregate_samples(input_file, output_dir = ".")
}
\arguments{
\item{input_file}{Path to file with interaction results for all samples (401_combine_samples/401_samples_interactions_agg_rank.rds)}

\item{output_dir}{output directory for saving output (default = '.')}
}
\description{
Combine p-values/scores by condition
}
