% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_combine_samples.R
\name{combine_samples}
\alias{combine_samples}
\title{Combine samples}
\usage{
combine_samples(
  input_dir,
  metadata,
  patient_var,
  condition_var = "Condition_dummy",
  sample_var = "Sample",
  output_dir = "."
)
}
\arguments{
\item{input_dir}{directory with files ending in '__interactions_mvoted.rds', '__signif_interactions.rds' and '*__interactions_agg_rank.rds',}

\item{metadata}{path to file with corresponding metadata (RDS)}

\item{patient_var}{patient variable name in metadata}

\item{condition_var}{condition variable name in metadata (default = "Condition_dummy")}

\item{sample_var}{sample variable name in metadata (default = "Sample")}

\item{output_dir}{output directory for saving output (default = '.')}
}
\description{
hard-combine results per sample.
}
