% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_in_multi_samples.R
\name{filter_by_detection_in_multi_samples}
\alias{filter_by_detection_in_multi_samples}
\title{Filter by detection in multiple samples}
\usage{
filter_by_detection_in_multi_samples(
  input_file,
  min_patients = 2,
  condition_var = "Condition_dummy",
  output_dir = "."
)
}
\arguments{
\item{input_file}{Path to '401_samples_interactions_mvoted.rds' file}

\item{min_patients}{Minimum number of patients for an interaction to be kept (default = 2)}

\item{condition_var}{Variabile containing the 'condition' (group/category) of the samples (default = "Condition_dummy")}

\item{output_dir}{output directory for saving output (default = '.')}
}
\description{
Filter by detection in multiple samples
}
