% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cell2cell.R
\name{format_cell2cell}
\alias{format_cell2cell}
\title{Format cell2cell CCI results}
\usage{
format_cell2cell(
  input_interactions,
  ref_db = "data/interactions_db/ref_db.rds",
  output_dir = ".",
  sample_id = NULL,
  mode = "pval"
)
}
\arguments{
\item{input_interactions}{path to file with the inferred interactions from cell2cell (output file generated by Python/202_cci_cell2cell.py, a csv file)}

\item{ref_db}{Path to interactions database (default = "data/interactions_db/ref_db.rds")}

\item{output_dir}{output directory for saving output (default = '.')}

\item{sample_id}{sample id to use for saving formatted CCI results (default = NULL; will determine sample id based on 'input_interactions' for this input_interactions has to be of the format 'cell2cell__{sample_id}.csv')}

\item{mode}{'pval' or 'Cell2Cell_score' dependent on input}
}
\description{
Format cell2cell CCI results
}
