% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cellchat.R
\name{format_cellchat}
\alias{format_cellchat}
\title{Format CellChat CCI results}
\usage{
format_cellchat(
  input_interactions,
  ref_db = "data/interactions_db/ref_db.rds",
  output_dir = ".",
  sample_id = NULL
)
}
\arguments{
\item{input_interactions}{path to file with the inferred interactions from CellChat (output file generated by run_cellchat()) (rds file)}

\item{ref_db}{Path to interactions database (default = "data/interactions_db/ref_db.rds")}

\item{output_dir}{output directory for saving output (default = '.')}

\item{sample_id}{sample id to use for saving formatted CCI results (default = NULL; will determine sample id based on 'input_interactions' for this input_interactions has to be of the format 'cellchat__{sample_id}.rds')}
}
\description{
Format CellChat CCI results
}
