% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cpdb.R
\name{format_cpdb}
\alias{format_cpdb}
\title{Format cpdb CCI results}
\usage{
format_cpdb(
  interaction_scores,
  pval,
  sign_means,
  means,
  ref_db = "data/interactions_db/ref_db.rds",
  output_dir = ".",
  sample_id = NULL
)
}
\arguments{
\item{interaction_scores}{path to 'statistical_analysis_interaction_scores' txt file}

\item{pval}{path to 'statistical_analysis_pvalues' txt file}

\item{sign_means}{path to 'statistical_analysis_significant_means' txt file}

\item{means}{path to 'statistical_analysis_means' txt file}

\item{ref_db}{Path to interactions database (default = "data/interactions_db/ref_db.rds")}

\item{output_dir}{output directory for saving output (default = '.')}

\item{sample_id}{sample id to use for saving formatted CCI results (default = NULL; will determine sample id based on 'input_interactions' for this input_interactions has to be of the format 'cpdb__{sample_id}.csv')}
}
\description{
Format cpdb CCI results
}
