% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellchat.R
\name{run_cellchat}
\alias{run_cellchat}
\title{Run CellChat}
\usage{
run_cellchat(
  gene_expr,
  annot,
  interactions_db,
  output_dir = ".",
  min_cells = 5,
  n_perm = 1000
)
}
\arguments{
\item{gene_expr}{seurat object with gene expression (rds file)}

\item{annot}{Column in metadata containing the cell type labels}

\item{interactions_db}{(custom) CellChat database}

\item{output_dir}{output directory for saving output (default = '.')}

\item{min_cells}{Minimum number of cells required in each cell group for cell-cell communication (default = 5)}

\item{n_perm}{Number of permutations for permutation testing (default = 1000)}
}
\description{
Run CellChat
}
