% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_methods.R
\name{take_consensus}
\alias{take_consensus}
\title{RRA interactions}
\usage{
take_consensus(
  cellchat_obj,
  liana_obj,
  cell2cell_obj,
  cpdb_obj,
  sample_id = NULL,
  output_dir = ".",
  alpha = 0.05
)
}
\arguments{
\item{cellchat_obj}{path to file generated by 'format_cellchat()'}

\item{liana_obj}{path to file generated by 'format_liana()'}

\item{cell2cell_obj}{path to file generated by 'format_cell2cell()'}

\item{cpdb_obj}{path to file generated by 'format_cpdb()'}

\item{sample_id}{sample id to use for saving object}

\item{output_dir}{output directory for saving output (default = '.')}

\item{alpha}{significance threshold (default = 0.05)}
}
\description{
Rank interactions based on CellChat, LIANA (use all methods), cell2cell and CellPhoneDB
}
